Base d'un projet opérationnel
Une version de base d'un projet qui fonctionne parfaitement.

+++ core/src/com/kamigaku/dungeoncrawler/component/GraphicsComponent.java
@@ -11,7 +11,7 @@ public class GraphicsComponent {
         this._sprite = sprite;
     }
     
-    public void update(SpriteBatch batch, int x, int y) {
+    public void update(SpriteBatch batch, float x, float y) {
         batch.draw(this._sprite.getTexture(), x, y);
     }
